[feat]: 타입 오류 해결 (#96)

+++ app/src/main/java/com/texthip/thip/data/service/FeedService.kt
@@ -98,7 +98,7 @@ interface FeedService {
     /** 피드 수정 */
     @PATCH("feeds/{feedId}")
     suspend fun updateFeed(
-        @Path("feedId") feedId: Int,
+        @Path("feedId") feedId: Long,
         @Body request: UpdateFeedRequest
     ): BaseResponse<CreateFeedResponse>
